Move imzml browser code from rest/api
